+++ src/java/net/jpountz/lz4/LZ4BlockInputStream.java
@@ -21,7 +21,6 @@ import static net.jpountz.lz4.LZ4BlockOutputStream.DEFAULT_SEED;
 import static net.jpountz.lz4.LZ4BlockOutputStream.HEADER_LENGTH;
 import static net.jpountz.lz4.LZ4BlockOutputStream.MAGIC;
 import static net.jpountz.lz4.LZ4BlockOutputStream.MAGIC_LENGTH;
-import static net.jpountz.lz4.LZ4BlockOutputStream.MIN_BLOCK_SIZE;
 
 import java.io.EOFException;
 import java.io.FilterInputStream;
@@ -36,12 +35,11 @@ import net.jpountz.xxhash.XXHashFactory;
 
 /**
  * {@link InputStream} implementation to decode data written with
- * {@link LZ4BlockOutputStream}. This class is not thread-safe.
+ * {@link LZ4BlockOutputStream}. This class is not thread-safe and does not
+ * support {@link #mark(int)}/{@link #reset()}.
  * @see LZ4BlockOutputStream
  */
 public final class LZ4BlockInputStream extends FilterInputStream {
-
-  private static final byte[] EMPTY = new byte[0];
 
   private final LZ4Decompressor decompressor;
   private final Checksum checksum;
@@ -50,11 +48,6 @@ public final class LZ4BlockInputStream extends FilterInputStream {
   private int originalLen;
   private int o;
   private boolean finished;
-
-  // for mark / reset
-  private byte[] markBuffer;
-  private int markOriginalLen, markO;
-  private boolean markFinished;
 
   /**
    * Create a new {@link InputStream}.
@@ -153,7 +146,7 @@ public final class LZ4BlockInputStream extends FilterInputStream {
   }
 
   private void refill() throws IOException {
-    in.read(compressedBuffer, 0, HEADER_LENGTH);
+    readFully(compressedBuffer, HEADER_LENGTH);
     for (int i = 0; i < MAGIC_LENGTH; ++i) {
       if (compressedBuffer[i] != MAGIC[i]) {
         throw new IOException("Stream is corrupted");
@@ -230,54 +223,17 @@ public final class LZ4BlockInputStream extends FilterInputStream {
 
   @Override
   public boolean markSupported() {
-    return in.markSupported();
+    return false;
   }
 
   @Override
   public void mark(int readlimit) {
-    readlimit = Math.max(readlimit, 0);
-    // worst-case compression ratio is when all blocks are MIN_BLOCK_SIZE bytes
-    // and incompressible
-    long compressedReadLimit =
-        (long) (MIN_BLOCK_SIZE + HEADER_LENGTH) // max compressed size of a single block
-        * (readlimit + MIN_BLOCK_SIZE - 1) / MIN_BLOCK_SIZE; // number of blocks
-    if (compressedReadLimit > Integer.MAX_VALUE) {
-      // overflow: try to do our best, this might not work if the input is incompressible
-      // and if the underlying InputStream actually buffers data when mark is called (if
-      // it does not support seeking)
-      compressedReadLimit = Integer.MAX_VALUE;
-    }
-    in.mark((int) compressedReadLimit);
-    markFinished = finished;
-    markO = o;
-    markOriginalLen = originalLen;
-    if (o < originalLen) {
-      if (markBuffer == null) {
-        markBuffer = new byte[originalLen];
-      } else if (markBuffer.length < originalLen - o) {
-        markBuffer = new byte[Math.max(markBuffer.length + markBuffer.length >>> 1, originalLen)];
-      }
-      System.arraycopy(buffer, o, markBuffer, 0, originalLen - o);
-    } else if (markBuffer == null) {
-      markBuffer = EMPTY;
-    }
-    assert markBuffer != null;
+    // unsupported
   }
 
   @Override
   public void reset() throws IOException {
-    if (markBuffer == null) {
-      throw new IOException("Call mark first");
-    }
-    in.reset();
-    finished = markFinished;
-    o = markO;
-    originalLen = markOriginalLen;
-    assert o <= originalLen;
-    if (o < originalLen) {
-      assert buffer.length >= originalLen; // block has alread been decompressed
-      System.arraycopy(markBuffer, 0, buffer, o, originalLen - o);
-    }
+    throw new IOException("mark/reset not supported");
   }
 
   @Override